test-files/python/requirements.txt
Appium-Python-Client==0.24
pytest==3.0.7
